requirements-dev.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
black==20.8b1
click==7.1.2
colorama==0.4.3
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==7.2.0
pylint==2.5.3
python-decouple==3.3
pytz==2020.1
regex==2020.11.13
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
